Polish apache/dubbo#5321 : Remove AnnotationUtils
